dev/be-diary

+++ server/wrieating/src/main/java/com/mainproject/wrieating/meal/entity/Day.java
@@ -8,12 +8,12 @@ import lombok.Setter;
 @Getter
 @NoArgsConstructor
 public class Day {
-    private Double totalKcal;
-    private Double totalCarbohydrate;
-    private Double totalProtein;
-    private Double totalFat;
+    private Long totalKcal;
+    private Long totalCarbohydrate;
+    private Long totalProtein;
+    private Long totalFat;
 
-    public Day(Double totalKcal, Double totalCarbohydrate, Double totalProtein, Double totalFat) {
+    public Day(Long totalKcal, Long totalCarbohydrate, Long totalProtein, Long totalFat) {
         this.totalKcal = totalKcal;
         this.totalCarbohydrate = totalCarbohydrate;
         this.totalProtein = totalProtein;
